Allow docker containers to verify tokens from keycloak
The default host name in keycloak is "localhost" this allows only
request that have localhost in them to verify any tokens.

Running a curl command from a docker container will result in the
following error;

{"error":"invalid_token","error_description":"Token invalid: Invalid token issuer. Expected 'http://keycloak:8080/auth/realms/development', but was 'http://localhost:8180/auth/realms/development'"}

To get around this I have set the hostname to be "keycloak" and
configured the port to be 8180 (As the port must match as well!)

The biggest change in the PR will be the keycloak port number will
be 8180 within docker. I can't see this affecting other teams as
we seem to be the first team to run into this issue.