Derive service/container name automatically
Update some of the common helper commands to be able to automatically
detect the service a user might be editing based on their working
directory.

This means that if you current working directory is somewhere inside,
for example, `index-map-api`, and you want to run its unit tests, you
can just type `unit-test` instead of `unit-test index-map-api`.